requirements.txt
customtkinter>=5.2.0
psutil>=5.8.0
Pillow>=10.0.0